LWDEV-5079 trades

+++ src/Lykke.Service.Operations.Contracts/AssetPairModel.cs
@@ -5,7 +5,7 @@
         public string Id { get; set; }
         public AssetModel BaseAsset { get; set; }
         public AssetModel QuotingAsset { get; set; }
-        public decimal MinVolume { get; set; }
-        public decimal MinInvertedVolume { get; set; }
+        public double MinVolume { get; set; }
+        public double MinInvertedVolume { get; set; }
     }
 }
